+++ test/unit/security-rules/security-rules-api-client.spec.ts
@@ -19,8 +19,8 @@
 import * as _ from 'lodash';
 import * as chai from 'chai';
 import * as sinon from 'sinon';
-import { SecurityRulesApiClient, RulesetContent } from '../../../src/security-rules/security-rules-internal';
-import { FirebaseSecurityRulesError } from '../../../src/security-rules/security-rules-utils-internal';
+import { SecurityRulesApiClient, RulesetContent } from '../../../src/security-rules/security-rules-api-client-internal';
+import { FirebaseSecurityRulesError } from '../../../src/security-rules/security-rules-internal';
 import { HttpClient } from '../../../src/utils/api-request';
 import * as utils from '../utils';
 import * as mocks from '../../resources/mocks';
